Add Connection header
https://www.w3.org/Protocols/rfc2616/rfc2616-sec14.html

HTTP/1.1 defines the "close" connection option for the sender to signal
that the connection will be closed after completion of the response. For
example,

  Connection: close

in either the request or the response header fields indicates that the
connection SHOULD NOT be considered `persistent' (secion 8.1) after the
current request/response is complete.

HTTP/1.1 applications that do not support persistent connections MUST
include "close" connection option in every message.